# Coursework/Databases/Backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Coffee Shop Backend API
  version: 1.0.0
  description: API локального backend-сервиса кофейни.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/shift/close:
    post:
      tags: [shift]
      summary: Закрыть активную смену
      responses:
        '200':
          description: Смена закрыта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloseShiftResponse'
        '400':
          $ref: '#/components/responses/ShiftNotActive'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/report/sales/excel:
    post:
      tags: [report]
      summary: Сформировать Excel-отчет по продажам
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeRangeRequest'
      responses:
        '200':
          $ref: '#/components/responses/ExcelFile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/report/menu-popularity/excel:
    post:
      tags: [report]
      summary: Сформировать Excel-отчет по популярности меню
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeRangeRequest'
      responses:
        '200':
          $ref: '#/components/responses/ExcelFile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
...
  responses:
    ExcelFile:
      description: Excel файл
      content:
        application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
          schema:
            type: string
            format: binary
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: invalid request body
...
    ShiftNotActive:
      description: Нет активной смены
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: SHIFT_NOT_ACTIVE
              message: Активная смена отсутствует
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: internal server error
  schemas:
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
    DutyEmployee:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        login:
          type: string
    Shift:
      type: object
      properties:
        id:
          type: string
        opened_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
        is_active:
          type: boolean
        duty:
          $ref: '#/components/schemas/DutyEmployee'
...
    CloseShiftResponse:
      type: object
      properties:
        shift:
          $ref: '#/components/schemas/Shift'
...
    TimeRangeRequest:
      type: object
      required: [from, to]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        employee_id:
          type: string
          nullable: true
...